chore: rename core to common
