+++ amman-common/src/amman-common.ts
