+++ amman-common/src/consts/index.ts
