+++ amman/package.json
@@ -52,7 +52,7 @@
     "@solana/spl-token": "^0.2.0",
     "@solana/spl-token-registry": "^0.2.2405",
     "@solana/web3.js": "^1.35.0",
-    "@metaplex-foundation/amman-core": "^0.0.0",
+    "@metaplex-foundation/amman-common": "^0.0.0",
     "ansi-colors": "^4.1.1",
     "bn.js": "^5.2.0",
     "buffer-hexdump": "^1.0.0",
