+++ package.json
@@ -1,13 +1,13 @@
 {
   "scripts": {
-    "build": "(cd ./amman-core && yarn build) && (cd ./amman && yarn build)",
+    "build": "(cd ./amman-common && yarn build) && (cd ./amman && yarn build)",
     "doc": "(cd ./amman && yarn doc);",
     "doc:update": "./sh/update-docs"
   },
   "workspaces": {
     "packages": [
       "amman",
-      "amman-core"
+      "amman-common"
     ]
   },
   "author": "Thorsten Lorenz <thlorenz@gmx.de>",
